requirements.txt
numpy==1.26.4
pandas==2.0.1
tifffile==2023.8.12
pystackreg==0.2.8
trackpy==0.7
scipy==1.11.1
matplotlib==3.7.2
torch>=2.0.0
torchvision>=0.15.0
tqdm
tensorboard>=2.14.0
Pillow>=9.0.0
